Draft: Upload attachment from clipboard
`:upload` argument is made optional. If omitted, then it will try to
grab an image from the clipboard instead.

The user might not know this behaviour and the command does not
indicate that the clipboard will be grabbed and uploaded without any
confirmation. There could be sensitive information in it.

There should be some mechanism to confirm/ask the user, or await a
"manual paste" (Ctrl-V? Shift-Ctrl-V?).